Clickable Buttons and Potential Posters
Made changes to main.ts using .mixin which allowed to direct the user to a link when clicking a certain button. These current links are filler, but can later be changed to different .vue files.

In GameCarousel.vue, a summary of changes include adding game from games.json, this path can be changed to pull games.json from MongoDB. Make sure the version of games.json is this commit's version, otherwise there will be errors. In the script portion I added a for loop that goes through posters for display, unfortunately the poster images will not display, but the alt names will.