updates to admin_accn_creation prototype
made updates to admin_accn_creation prototype that is close to being able to display all users within our specific firebase in the console of our project, soon to be updated to the admin page. Current errors with admin_accn_creation can be located in the console section when inspecting element after running dev.